requirements.txt
Django==1.8.2
channels==0.1.1
cssselect==0.9.1
dj-database-url==0.3.0
lxml==3.4.4
jsonfield==1.0.3
psycopg2==2.6.1
requests==2.7.0
djangorestframework==3.1.3
django-extensions==1.5.5
